split API functions
